[DEV-234] test TxAcceptedInBlock and TxBlocks

+++ blockdag/test_utils.go
@@ -49,53 +49,56 @@ func fileExists(name string) bool {
 // DAGSetup is used to create a new db and chain instance with the genesis
 // block already inserted.  In addition to the new chain instance, it returns
 // a teardown function the caller should invoke when done testing to clean up.
-func DAGSetup(dbName string, params *dagconfig.Params) (*BlockDAG, func(), error) {
+func DAGSetup(dbName string, params *dagconfig.Params, config Config) (*BlockDAG, func(), error) {
 	if !isSupportedDbType(testDbType) {
 		return nil, nil, fmt.Errorf("unsupported db type %v", testDbType)
 	}
 
-	// Handle memory database specially since it doesn't need the disk
-	// specific handling.
-	var db database.DB
 	var teardown func()
-	// Create the root directory for test databases.
-	if !fileExists(testDbRoot) {
-		if err := os.MkdirAll(testDbRoot, 0700); err != nil {
-			err := fmt.Errorf("unable to create test db "+
-				"root: %v", err)
-			return nil, nil, err
+
+	if config.DB == nil {
+		// Handle memory database specially since it doesn't need the disk
+		// specific handling.
+		var db database.DB
+		// Create the root directory for test databases.
+		if !fileExists(testDbRoot) {
+			if err := os.MkdirAll(testDbRoot, 0700); err != nil {
+				err := fmt.Errorf("unable to create test db "+
+					"root: %v", err)
+				return nil, nil, err
+			}
 		}
-	}
 
-	// Create a new database to store the accepted blocks into.
-	dbPath := filepath.Join(testDbRoot, dbName)
-	_ = os.RemoveAll(dbPath)
-	ndb, err := database.Create(testDbType, dbPath, blockDataNet)
-	if err != nil {
-		return nil, nil, fmt.Errorf("error creating db: %v", err)
-	}
-	db = ndb
+		// Create a new database to store the accepted blocks into.
+		dbPath := filepath.Join(testDbRoot, dbName)
+		_ = os.RemoveAll(dbPath)
+		ndb, err := database.Create(testDbType, dbPath, blockDataNet)
+		if err != nil {
+			return nil, nil, fmt.Errorf("error creating db: %v", err)
+		}
+		db = ndb
 
-	// Setup a teardown function for cleaning up.  This function is
-	// returned to the caller to be invoked when it is done testing.
-	teardown = func() {
-		db.Close()
-		os.RemoveAll(dbPath)
-		os.RemoveAll(testDbRoot)
+		config.DB = db
+
+		// Setup a teardown function for cleaning up.  This function is
+		// returned to the caller to be invoked when it is done testing.
+		teardown = func() {
+			db.Close()
+			os.RemoveAll(dbPath)
+			os.RemoveAll(testDbRoot)
+		}
 	}
 
 	// Copy the chain params to ensure any modifications the tests do to
 	// the chain parameters do not affect the global instance.
 	paramsCopy := *params
 
+	config.DAGParams = &paramsCopy
+	config.TimeSource = NewMedianTime()
+	config.SigCache = txscript.NewSigCache(1000)
+
 	// Create the DAG instance.
-	dag, err := New(&Config{
-		DB:          db,
-		DAGParams:   &paramsCopy,
-		Checkpoints: nil,
-		TimeSource:  NewMedianTime(),
-		SigCache:    txscript.NewSigCache(1000),
-	})
+	dag, err := New(&config)
 	if err != nil {
 		teardown()
 		err := fmt.Errorf("failed to create dag instance: %v", err)
